Only  kill the pid if the pid file exists exists

+++ src/saltbox/salt_helpers.py
@@ -75,8 +75,8 @@ class SaltDaemon:
         if not self.running:
             return
         pid = self.pid
-        assert pid is not None
-        os.kill(pid, self.SIGHUP_SIGNAL)
+        if pid is not None:
+            os.kill(pid, self.SIGHUP_SIGNAL)
         self._running = False
 
     def wait(self):
@@ -95,6 +95,8 @@ class SaltDaemon:
     @property
     def pid(self):
         pid_file_path = os.path.join(self._config_obj.salt_run_path, self.PIDFILE)
+        if not os.path.exists(pid_file_path):
+            return None
         return int(open(pid_file_path).read())
 
     # def __enter__(self):
